add tcp tracer
