+++ cmd/aks-periscope/aks-periscope.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/Azure/aks-periscope/pkg/collector"
+	"github.com/Azure/aks-periscope/pkg/collector/inspektor-gadget"
 	"github.com/Azure/aks-periscope/pkg/diagnoser"
 	"github.com/Azure/aks-periscope/pkg/exporter"
 	"github.com/Azure/aks-periscope/pkg/interfaces"
@@ -98,7 +99,8 @@ func run(osIdentifier utils.OSIdentifier, knownFilePaths *utils.KnownFilePaths, 
 		collector.NewSystemLogsCollector(osIdentifier, runtimeInfo),
 		collector.NewSystemPerfCollector(config, runtimeInfo),
 		collector.NewWindowsLogsCollector(osIdentifier, runtimeInfo, knownFilePaths, fileSystem, 10*time.Second, 20*time.Minute),
-		collector.NewInspektorGadgetDNSTraceCollector(config, runtimeInfo),
+		inspektor_gadget.NewInspektorGadgetDNSTraceCollector(config, runtimeInfo),
+		inspektor_gadget.NewInspektorGadgetTCPTraceCollector(config, runtimeInfo),
 	}
 
 	collectorGrp := new(sync.WaitGroup)
