+++ pkg/test/resources/tools-resources/chaos/coredns.yaml
@@ -1,0 +1,16 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: coredns-custom
+  namespace: kube-system
+  annotations:
+    kubectl.kubernetes.io/last-applied-configuration: poos and wees
+data:
+  yourdns.server: |
+    .:53 {
+        health
+    
+        . {
+            whoami
+        }
+    }
