+++ pkg/test/resources/tools-resources/chaos/readme.md
@@ -1,0 +1,3 @@
+# Chaos 
+
+The folder contains configuration that can break your cluster.
