+++ pkg/test/resources/tools-resources/inspektor-gadget/tcp-trace.yaml
@@ -1,0 +1,12 @@
+apiVersion: gadget.kinvolk.io/v1alpha1
+kind: Trace
+metadata:
+  name: tcptracer-t27aa
+  namespace: gadget
+  annotations:
+    gadget.kinvolk.io/operation: start
+spec:
+  gadget: tcptracer
+  runMode: Manual
+  node: aks-nodepool1-11865456-vmss000000
+  outputMode: Stream
